requirements.txt
wheel>=0.40.0
setuptools>=68.0.0
cairocffi>=0.9.0
cairosvg==2.7.1
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.1
Pillow
streamlit>=1.28.2
requests>=2.31.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.3.5
pydantic